API
전체 채팅목록 가져오기 (메인화면),
채팅 추가하기

+++ app/src/main/java/com/chat_soon_e/re_chat/data/entities/FolderContent.kt
@@ -8,7 +8,7 @@ import java.io.Serializable
 @Entity(tableName = "FolderContentTable")
 data class FolderContent(
     var folderIdx: Int,
-    var chatIdx: Int,
+    var chatIdx: Long,
     var status: String? = "ACTIVE"   // ACTIVE, DELETED
 ):Serializable{
     @PrimaryKey(autoGenerate = true) var folderContentIdx: Int = 0
